requirements.txt
streamlit
textblob
python-docx